code/tester_code/requirements.txt
pyserial==3.3
pyfiglet
inputs
keyboard